+++ tango/common/__init__.py
@@ -1,6 +1,22 @@
-from .dataset_dict import DatasetDict
+from .aliases import PathOrStr
+from .dataset_dict import DatasetDictBase, DatasetDict, IterableDatasetDict
+from ._det_hash import CustomDetHash, det_hash
 from .params import Params
 from .from_params import FromParams
 from .registrable import Registrable
 from .lazy import Lazy
 from .tqdm import Tqdm
+
+__all__ = [
+    "PathOrStr",
+    "DatasetDictBase",
+    "DatasetDict",
+    "IterableDatasetDict",
+    "CustomDetHash",
+    "det_hash",
+    "Params",
+    "FromParams",
+    "Registrable",
+    "Lazy",
+    "Tqdm",
+]